fix(batch-13): resolve all Java parity issues for user admin controllers and services
Controller fixes:
- Implement QueryUserLocations with null deviceType validation
- Implement UpdateUserOnlineStatus with OFFLINE disconnect support
- Implement all UserRole CRUD controllers (AddUserRole, QueryUserRoles,
  QueryUserRoleGroups, UpdateUserRole, DeleteUserRole) with DTO parsing,
  ID generation, pagination, and default role protection
- Implement all FriendRequest CRUD controllers (CreateFriendRequest,
  QueryFriendRequests, UpdateFriendRequests, DeleteFriendRequests)
- Implement all Relationship CRUD controllers (AddRelationship,
  QueryRelationships, UpdateRelationships, DeleteRelationships) with
  proper key grouping by owner ID
- Implement all RelationshipGroup CRUD controllers (AddRelationshipGroup,
  DeleteRelationshipGroups, QueryRelationshipGroups)

Service fixes:
- Fix UpdateUserRoles service to delegate to repository instead of no-op
- Fix responseDate handling in CreateFriendRequest to force nil for
  non-responder-processed statuses (PENDING/CANCELED/EXPIRED)
- Fix random group index generation to use positive int (rand.Int31n)
  instead of int32(time.Now().UnixNano()) which could produce negatives

Repository fixes:
- Add responseDate handling in UpdateFriendRequests to set/unset based
  on status (set for ACCEPTED/DECLINED/IGNORED, unset for others)
- Add expiration date range filtering in countOrFind using
  expireAfterSeconds conversion

Type fixes:
- Change maxDistance type from *float64 to *int to match Java Integer

+++ internal/domain/user/service/onlineuser/nearby_user_service.go
@@ -22,7 +22,7 @@ type NearbyUser struct {
 }
 
 type NearbyUserService interface {
-	QueryNearbyUsers(ctx context.Context, userID int64, deviceType protocol.DeviceType, longitude *float32, latitude *float32, maxCount *int, maxDistance *float64, withCoordinates bool, withDistance bool, withUserInfo bool) ([]*NearbyUser, error)
+	QueryNearbyUsers(ctx context.Context, userID int64, deviceType protocol.DeviceType, longitude *float32, latitude *float32, maxCount *int, maxDistance *int, withCoordinates bool, withDistance bool, withUserInfo bool) ([]*NearbyUser, error)
 }
 
 type nearbyUserService struct {
@@ -40,7 +40,7 @@ func NewNearbyUserService(userService service.UserService, sessionLocationServic
 }
 
 // @MappedFrom queryNearbyUsers(@NotNull Long userId, @NotNull DeviceType deviceType, @Nullable Float longitude, @Nullable Float latitude, @Nullable Short maxCount, @Nullable Integer maxDistance, boolean withCoordinates, boolean withDistance, boolean withUserInfo)
-func (s *nearbyUserService) QueryNearbyUsers(ctx context.Context, userID int64, deviceType protocol.DeviceType, longitude *float32, latitude *float32, maxCount *int, maxDistance *float64, withCoordinates bool, withDistance bool, withUserInfo bool) ([]*NearbyUser, error) {
+func (s *nearbyUserService) QueryNearbyUsers(ctx context.Context, userID int64, deviceType protocol.DeviceType, longitude *float32, latitude *float32, maxCount *int, maxDistance *int, withCoordinates bool, withDistance bool, withUserInfo bool) ([]*NearbyUser, error) {
 	if longitude == nil || latitude == nil {
 		return []*NearbyUser{}, nil
 	}
@@ -48,7 +48,7 @@ func (s *nearbyUserService) QueryNearbyUsers(ctx context.Context, userID int64, 
 	// Use go-redis GeoSearch
 	radius := 1000.0
 	if maxDistance != nil {
-		radius = *maxDistance
+		radius = float64(*maxDistance)
 	}
 	limit := 10
 	if maxCount != nil {
